Added NMultiDrawable, NPaddedDrawable, and NTileImageDrawable.
The only one that probably needs immediate explanation is MultiDrawable, which is simply a drawable that draws other drawables.  You pass it an iterable, it iterates over the drawables in the iterable you give it and draws them, that's it. Nothing incredibly special, but it can be handy for building drawables that are slightly more complex than the average drawable, especially when you consider that some controls (like buttons or windows) only accept a single drawable for their main form, when you might want to have more than one.