Removing NGUI's singleton status.  In doing so, I am breaking any code that previously worked, including most of NUIT.  Thankfully, this commit includes those fixes as well!  O frabjous day!  Callooh! Callay!  This means makeActive, makeInactive, activeInstance, __instance, etc. are gone.  In fact, the init routine is gone as well, because there's no need for it to exist now.
In other news, NView has gained _drawable and _font members, the use of which should probably be clear.  NGUI has a viewFont NFont member now that is considered the default font for all views unless you specify it should use another font.  As a rule of thumb, view subclasses should probably not have fonts associated with them because there's no way to ensure distribution of a given font, nor is there a way to guarantee a given renderer can load the provided font.

NWindow and NPopup now use ndd's nice 'super func' idiom, which is solely responsible for increasing 14.2% of the world's happiness, for their init routines.